Move more logic to main
